+++ src/test/suite/kafka-file/languageservice/completionProperties.test.ts
@@ -567,7 +567,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 0))
                     },
                     {
@@ -613,7 +613,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 1))
                     },
                     {
@@ -658,7 +658,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 5))
                     },
                     {
@@ -703,7 +703,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
                     },
                     {
@@ -748,7 +748,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 6))
                     },
                     {
@@ -793,7 +793,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(1, 0), position(1, 11))
                     },
                     {
@@ -840,7 +840,7 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     },
                     {
                         label: 'value-format', kind: CompletionItemKind.Property,
-                        insertText: 'value-format: ${1|string,json,double,float,integer,long,short|}',
+                        insertText: 'value-format: ${1|string,json,avro,double,float,integer,long,short|}',
                         range: range(position(2, 0), position(2, 5))
                     },
                     {
@@ -1059,6 +1059,11 @@ suite("Kafka File PRODUCER Completion Test Suite", () => {
                     {
                         label: 'json', kind: CompletionItemKind.Value,
                         insertText: ' json',
+                        range: range(position(1, 13), position(1, 13))
+                    },
+                    {
+                        label: 'avro', kind: CompletionItemKind.Value,
+                        insertText: ' avro',
                         range: range(position(1, 13), position(1, 13))
                     },
                     {